object_detection/chanan/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.3
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
lvis==0.5.3
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.33.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0+cu126
torchaudio==2.8.0+cu126
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.0
typing_extensions==4.15.0
ultralytics==8.3.191
ultralytics-thop==2.0.16
urllib3==2.5.0
wheel==0.45.1
